Do not memoize selected sap profile
SAP Hana profiles have pre-set processor and memory values. The
provisioning form should dynamically update the values based on the
selected profile (which is selected through the "Machine Type" field,
when the provisioning image is a SAP Hana image).

When the selected flavor is memoized the UI never retrieved the newly
selected profile to update processor and memory values accordingly.