Use consistent value types for 'affinity_policy'
The Power Cloud API new volume create call
(https://cloud.ibm.com/apidocs/power-cloud#pcloud-cloudinstances-volumes-post)
can set `affinityPolicy` to only 'affinity' or 'anti-affinity'. When no
affinity policy is required the 'affinityPolicy' is omitted (or set to
nil).

I attempted to implement this directly in params_for_create ddf
definition, but for some reason the 'affinity_policy' :label was set in
the resulting options instead of its :value. This caused the API call to
send '"affinityPolicy": "Off"', which is invalid.

This fix sets the off :value to 'Off' (same as the :label), both to be
consistent and to eliminate the need to distinguish between :label and
:value. That change requires additional logic in the 'raw_create_volume'
method in order to handle the 'Off' case correctly.