Set default endpoint in COS ems constructor
A default 'Endpoint' object is created in the COS ems 'endpoints'
attribute when the ems object is created. The '<<' operator will append
an _additional_ endpoint to the 'endpoints' list. Setting the url in the
default endpoint should be done through the 'FactorBot.create()' call to
create the COS ems object, so there is only one default endpoint with
the url set in this commit.