Fix PowerVS 'cpu_type' hardware parameter
When a PowerVS VM is build from a "stock" image, there is a transparent
copy made with a new image ID. This copy is _not_ included in the return
from the 'images' collector method, but can be accessed directly. The
'architecture' method added in this commits allows for the extra lookup
for this case, ensuring VMs set the 'cpu_type' parameter.