Eliminate unnecessary image API calls
The PowerVirtualServers collector 'images' method, which gets a list of
all images, contains all of the attributes used in the parser. The Power
Cloud API call for a single _image_ does contain additional attributes,
but they are not used in the MIQ inventory. We can reduce the number of
API calls by only making single image API calls if they're note included
in the get all 'images' list.