Move Stream sleep to before yielding events
PowerVS back-end can be kind of slow. The events are generated when a
state change is requested, resulting in some delay before a new state is
actually reported. When testing MIQ would usually refresh the VM before
the API reported a new state.

By moving the polling sleep time in-between getting the events and
yielding them to the event catcher there is a guaranteed (and
configurable) delay.