Add implicit :number_of_vms = 1 to dialog
There is logic in Automate that adds a suffix to VM names if more than
one is being provisioned[1]. The ':number_of_vms' is required to be set.
In this case, provisioning a template from an existing VM, the number of
VMs should always be 1. This can be set with a hidden field in the
dialog.

[1]: https://github.com/ManageIQ/manageiq-content/blob/master/content/automate/ManageIQ/Cloud/VM/Provisioning/Naming.class/__methods__/vmname.rb#L63