Fix SSH key fingerprint when newline(s) in comment
PowerVS allows newlines within public SSH key comments, but this causes
an exception when calculating the fingerprint. Since a comment does not
affect the fingerprint we can safely ignore any newline (and anything
afterwards) when computing the fingerprint. There's no problem with
storing the public key + comments as-is in the database where the
information may be useful in reports.