Fix AWS S3 VCR errors getting bucket location
Newer versions of the aws-s3 gem change how they call for the bucket
location constraing and the overly broad `rescue` to skip the bucket was
hiding the VCR unknown http interaction exception, leaving only the
unused http interactions at the end of the run.

This fixes the request URI for the bucket calls and adds a more specific
`Aws::S3::Errors::NoSuchBucket` exception check for skipping buckets not
in our region.