Use ems_ref, not name, to find new VM or Template
PowerVS back-end does require VM and template (boot image) names to be
unique within a provider (workspace) so for simple provisioning it does
work to use the name to find the newly created VM or template. When
provisioning from service catalogs, however, the 'options[:vm_name]' is
enumerated (four digits with leading zeroes) without changing the
'options[:vm_name]'. This causes the service request to fail even though
the VM was successfully created.