Fix clone_to_template provision preparation
The switch statement in create_destination caused clone_to_template to
skip the prepare_provision phase. Because of this, clone_options was set
in a start_clone_task override. This works perfectly for
clone_to_template, but when provisioning a vm *from* a template the
clone_options parsing gets calls twice - from prepare_provision and the
start_clone_task override.

The start_clone_task override can be dropped entirely if we simply call
prepare_provision for the clone_to_template case.