Add sleep to PowerVS vcr setup
After creating a PVM instance, the
pvm_instances_api.pcloud_pvminstances_get() for the new instance returns
an HTTP 500 response for a short time. It takes a few minutes for the
instance to be created anyways, so adding a 1m sleep prevents the bad
response without slowing down the setup process.